Clean up nimble files
